==> 04B_analyze_probabilistic_micro.Rmd <==
---
title: "Analysis of HB CE probabilistic model"
author: "Niklaus Meier"
date: "`r format(Sys.Date(), '%B %d, %Y')`"
output: 
  word_document:
    fig_caption: yes
    fig_height: 6
    fig_width: 8
    highlight: null
    reference_docx: word_template.docx
    toc: no
    toc_depth: 4
  # html_document:
  #   toc: true
  #   toc_float: true
  #   number_sections: true
editor_options: 
  chunk_output_type: console
---

<!-- Setup -->

```{r setup, include = FALSE}

knitr::opts_chunk$set(echo = TRUE)

#===============================================================================
# Load data
#===============================================================================

load(file = paste0(directories$dir_dat_deriv, "/probabilistic_model_results.Rdata"))
load(file = paste0(directories$dir_dat_deriv, "/probabilistic_model_diagnostics.Rdata"))


```

```{r Summary Tables, echo=FALSE}

summary_outcomes <- c("Treatment","LYs", "QALYs", "Costs", "NMB", "ICER")

## Comparison between discounted outcomes of ETRANACOGENE and Prophylaxis

summary_table_disc_ETRANACOGENE_PROPHYLAXIS <- setNames(data.frame(matrix(data = 0,
                                                       nrow = 3,
                                                       ncol = length(summary_outcomes))), 
                                                            summary_outcomes)

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[,"Treatment"] <- c("ETRANACOGENE", "Prophylaxis", "Increment")

summary_table_disc_ETRANACOGENE_PROPHYLAXIS <- setNames(data.frame(matrix(data = 0,
                                                       nrow = 3,
                                                       ncol = length(summary_outcomes))), 
                                                            summary_outcomes)

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[,"Treatment"] <- c("ETRANACOGENE", "Prophylaxis", "Increment")


summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"LYs"]      <- mean(model_results_combined[,"ETRANACOGENE_LYs_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"LYs"]      <- mean(model_results_combined[,"PROPHYLAXIS_LYs_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"LYs"]      <- summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"LYs"] -                                                                                           summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"LYs"]

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"QALYs"]    <- mean(model_results_combined[,"ETRANACOGENE_QALYs_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"QALYs"]    <- mean(model_results_combined[,"PROPHYLAXIS_QALYs_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"QALYs"]    <- summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"QALYs"] -                                                                                         summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"QALYs"]

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"Costs"]    <- mean(model_results_combined[,"ETRANACOGENE_cost_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"Costs"]    <- mean(model_results_combined[,"PROPHYLAXIS_cost_disc"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"Costs"]    <- summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"Costs"] -                                                                                         summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"Costs"]

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"NMB"]      <- mean(model_results_combined[,"ETRANACOGENE_NMB"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"NMB"]      <- mean(model_results_combined[,"PROPHYLAXIS_NMB"])
summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"NMB"]      <- summary_table_disc_ETRANACOGENE_PROPHYLAXIS[1,"NMB"] -                                                                                            summary_table_disc_ETRANACOGENE_PROPHYLAXIS[2,"NMB"]

summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"ICER"]     <- summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"Costs"] /                                                                                          summary_table_disc_ETRANACOGENE_PROPHYLAXIS[3,"QALYs"]

kbl(summary_table_disc_ETRANACOGENE_PROPHYLAXIS, format = "simple", digits = 2, caption = "Comparison discounted outcomes between ETRANACOGENE and Prophylaxis")

```

\newpage

```{r Diagnostics 1, include = FALSE}

model_diag$plots <- fun_diagnostics_plots(model_results_combined)


```

```{r Diagnostics 2, echo=FALSE}

## model diagnostics

for (i in 1:10) {
  print(ggarrange(plotlist = model_diag$plots[(1+(i-1)*4):(4+(i-1)*4)],
                labels = colnames(model_diag$convergence)[(1+(i-1)*4):(4+(i-1)*4)],
                ncol = 2, nrow = 2,
                font.label = list(size = 10, color = "dark blue", face = "bold")))
}


convergence_table <- data.frame(matrix(nrow = nrow(model_diag[["convergence"]]), ncol = 3))
colnames(convergence_table) <- c("ETRA vs. PROPH inc. QALYs", 
                                 "ETRA vs. PROPH inc. Costs", 
                                 "ETRA vs. PROPH inc. NMB")

convergence_table[,"ETRA vs. PROPH inc. QALYs"] <- model_diag[["convergence"]][,"ETRANACOGENE_PROPHYLAXIS_inc_QALY"]
convergence_table[,"ETRA vs. PROPH inc. Costs"] <- model_diag[["convergence"]][,"ETRANACOGENE_PROPHYLAXIS_inc_Costs"]
convergence_table[,"ETRA vs. PROPH inc. NMB"]   <- model_diag[["convergence"]][,"ETRANACOGENE_PROPHYLAXIS_INMB"]

print("We used intervals of 1000 patients, and checked whether the running mean of the incremental outcomes change by more than 1%.")
print("A 0 means that we have reached a stable value, i.e. we have achieved convergence and have a sufficient sample size, using this interval and this percentage change as a measure of sufficient stability.")
kbl(convergence_table, format = "simple", digits = 2, caption = "Diagnostics of convergence of running incremental outcomes over all patients in simulation")


```

\newpage

```{r Comparisons, echo=FALSE}

# Set colors and add more colors to this palette
color_brew <- colorRampPalette(brewer.pal(4, "Set1") )(4)

# Scatterplots of incremental QALYs and Costs
ETRA <- "ETRANACOGENE"
PROPH <- "Prophylaxis"

ggplot() +
  geom_point(data = model_results_combined, aes(x = ETRANACOGENE_PROPHYLAXIS_inc_QALY , y = ETRANACOGENE_PROPHYLAXIS_inc_Costs/1000000, color = PROPH), size = 0.8) +
  geom_point(data = model_results_combined, aes(x = mean(ETRANACOGENE_PROPHYLAXIS_inc_QALY) , y = mean(ETRANACOGENE_PROPHYLAXIS_inc_Costs/1000000)), size = 3, shape = 18) +
  scale_y_continuous(limits=c(-4,1),breaks=seq(-4,1,1)) +
  scale_x_continuous(limits=c(-1,3),breaks=seq(-1,3,1)) +
  geom_hline(yintercept = 0) +
  geom_vline(xintercept = 0) +
  ylab("Cost Difference [million EUR]") + xlab("QALY Difference") +
  settings$ggplot_theme +
  scale_color_manual(values = c("Prophylaxis"=color_brew[2])) +
  theme(axis.title.x = element_text(size = 16),
        axis.text.x = element_text(size = 14),
        axis.title.y = element_text(size = 16),
        legend.position= "none")


## Comparison between discounted outcomes of ETRANACOGENE and FIX Prophylaxis

summary_costs <- c("Treatment","ETRANACOGENE", "FIX Prophylaxis", "FIX for treatment of bleeds", "Hospitalization", "Surgery", "Total")

summary_table_costs <- setNames(data.frame(matrix(data = 0,
                                                  nrow = 2,
                                                  ncol = length(summary_costs))), 
                                                         summary_costs)

summary_table_costs[,"Treatment"] <- c("ETRANACOGENE", "FIX Prophylaxis")

summary_table_costs[1,"ETRANACOGENE"]                <- mean(model_results_combined[,"ETRANACOGENE_cost_ETRANACOGENE_disc"])
summary_table_costs[2,"ETRANACOGENE"]                <- mean(model_results_combined[,"PROPHYLAXIS_cost_ETRANACOGENE_disc"])

summary_table_costs[1,"FIX Prophylaxis"]             <- mean(model_results_combined[,"ETRANACOGENE_cost_PROPHYLAXIS_disc"])
summary_table_costs[2,"FIX Prophylaxis"]             <- mean(model_results_combined[,"PROPHYLAXIS_cost_PROPHYLAXIS_disc"])
  
summary_table_costs[1,"FIX for treatment of bleeds"] <- mean(model_results_combined[,"ETRANACOGENE_cost_factor_treat_bleed_disc"])
summary_table_costs[2,"FIX for treatment of bleeds"] <- mean(model_results_combined[,"PROPHYLAXIS_cost_factor_treat_bleed_disc"])
  
summary_table_costs[1,"Hospitalization"]             <- mean(model_results_combined[,"ETRANACOGENE_cost_hospitalization_disc"])
summary_table_costs[2,"Hospitalization"]             <- mean(model_results_combined[,"PROPHYLAXIS_cost_hospitalization_disc"])
  
summary_table_costs[1,"Surgery"]                     <- mean(model_results_combined[,"ETRANACOGENE_cost_surgery_disc"])
summary_table_costs[2,"Surgery"]                     <- mean(model_results_combined[,"PROPHYLAXIS_cost_surgery_disc"])
  
summary_table_costs[1,"Total"]                       <- sum(summary_table_costs[1, 2:ncol(summary_table_costs)])
summary_table_costs[2,"Total"]                       <- sum(summary_table_costs[2, 2:ncol(summary_table_costs)])

kbl(summary_table_costs, format = "simple", digits = 2, caption = "Discounted costs broken down by category")


```

\newpage

```{r Dominance, echo=FALSE}

dominant <- subset(model_results_combined, model_results_combined[,"ETRANACOGENE_PROPHYLAXIS_inc_Costs"] < 0)
dominant <- subset(dominant, dominant[,"ETRANACOGENE_PROPHYLAXIS_inc_QALY"] > 0)

share_dominance <- nrow(dominant)/nrow(model_results_combined)
print(paste0(share_dominance * 100, "% of all probabilistic simulations were dominant in favor of ED."))


```

\newpage

```{r Cost-effectiveness acceptability curve (CEAC), echo=FALSE}

# We recalculate the probability of therapies being cost-effective based on a shift in the willingness-to-pay

columns <- c("WTP", "ED","EHLFIX")

WTP <- seq(0,300000,5000)

table_CEAC <- setNames(data.frame(matrix(data = 0,
                                         nrow = length(WTP),
                                         ncol = length(columns))), 
                                                       columns)
table_CEAC[,"WTP"] <- WTP

for (i in 1:length(WTP)) {
  
  # Recalculate NMB based on new WTP
  NMB_ED <- model_results_combined[,"ETRANACOGENE_QALYs_disc"]*WTP[i] - model_results_combined[,"ETRANACOGENE_cost_disc"]
  NMB_EHLFIX <- model_results_combined[,"PROPHYLAXIS_QALYs_disc"]*WTP[i] - model_results_combined[,"PROPHYLAXIS_cost_disc"]
  
  # Count in how many simulations ED is or is not cost-effective
  table_CEAC[i,"ED"] <- sum((NMB_ED > NMB_EHLFIX) == 1) / nrow(model_results_combined)
  
  table_CEAC[i,"EHLFIX"] <- sum((NMB_ED < NMB_EHLFIX) == 1) / nrow(model_results_combined)
  
}

ggplot(table_CEAC, aes(x=WTP)) + 
  geom_line(aes(y = ED, color = ETRA), size = 1) + 
  geom_line(aes(y = EHLFIX, color = PROPH), size = 1) +
  ylab("Probability of being cost-effective") + xlab("Willingness-to-pay [EUR]") +
  settings$ggplot_theme +
  theme(legend.position = "bottom") +
  scale_color_manual(values = c("Prophylaxis"=color_brew[2],"ETRANACOGENE"=color_brew[1]))


```


# System information

```{r sysinfo, echo = TRUE, results = 'asis'}

Sys.info()

sessionInfo(package = NULL)

```

```{r remove, include = FALSE}

rm(model_results, model_diagnostics,
   color_brew, summary_outcomes)

gc()

```
